rename: to weather-current
